requirements.txt
Flask
streamlit
opencv-python
Pillow
requests
numpy
